backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
bs4==0.0.2
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.3
defusedxml==0.7.1
Django==5.2.1
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-yasg==1.21.10
exceptiongroup==1.3.0
flower==2.0.1
h11==0.16.0
humanize==4.12.3
idna==3.10
inflection==0.5.1
kombu==5.5.4
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==25.0
pillow==11.2.1
prometheus_client==0.22.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9
pycparser==2.22
PyJWT==2.9.0
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
selenium==4.33.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
tornado==6.5.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
undetected-chromedriver==3.5.5
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
